reseauneurone/requirements.txt
numpy
matplotlib
keras
tensorflow
pillow 
opencv-python
torch
transformers
openai